rename filter and filterlist classes for anime,
update libversion for anime extensions to 10

+++ app/src/main/java/eu/kanade/tachiyomi/extension/api/AnimeExtensionGithubApi.kt
@@ -59,7 +59,7 @@ internal class AnimeExtensionGithubApi {
         return json
             .filter { element ->
                 val versionName = element.jsonObject["version"]!!.jsonPrimitive.content
-                val libVersion = versionName.substringBeforeLast('.').toDouble()
+                val libVersion = versionName.substringBeforeLast('.').toInt()
                 libVersion >= AnimeExtensionLoader.LIB_VERSION_MIN && libVersion <= AnimeExtensionLoader.LIB_VERSION_MAX
             }
             .map { element ->
